requirements.txt
streamlit
python-dotenv
uuid
openai>=1.0.0
langchain>=0.1.14
langchain-community>=0.0.32
sentence-transformers
chromadb>=0.4.22
pypdf
protobuf<4.24.0
torch<2.2

azure-identity
azure-storage-blob
